requirements.txt
streamlit==1.52.2
openai-agents[redis]==0.6.5
upstash-vector==0.8.0
pytest==9.0.2
python-dotenv==1.2.1
pip install -r requirements.txt
